Backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
websockets==14.1
httptools==0.6.4
uvloop==0.21.0; sys_platform != 'win32'
watchfiles==1.0.3
supabase==2.10.0
python-dotenv==1.0.1
bcrypt==4.2.1
PyJWT==2.9.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20
email-validator==2.2.0
